+++ flutter_mobx/test/helpers.dart
@@ -2,7 +2,7 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:mobx/mobx.dart';
 import 'package:mobx/src/core.dart' show ReactionImpl;
-import 'package:mockito/mockito.dart';
+import 'package:mocktail/mocktail.dart';
 
 // ignore: top_level_function_literal_block, prefer_function_declarations_over_variables
 final voidFn = () {};
@@ -16,7 +16,7 @@ class MockReaction extends Mock implements ReactionImpl {
 
   @override
   bool get hasObservables =>
-      super.noSuchMethod(Invocation.getter(#hasObservables), returnValue: true);
+      super.noSuchMethod(Invocation.getter(#hasObservables));
 }
 
 // ignore: must_be_immutable