+++ mobx/test/autorun_test.dart
@@ -1,11 +1,13 @@
 import 'package:fake_async/fake_async.dart';
 import 'package:mobx/mobx.dart';
 import 'package:mobx/src/api/reaction.dart';
-import 'package:mockito/mockito.dart';
+import 'package:mocktail/mocktail.dart' as mock;
 import 'package:test/test.dart';
 
 import 'shared_mocks.dart';
 import 'util.dart';
+
+// ignore_for_file: unnecessary_lambdas
 
 void main() {
   testSetup();
@@ -132,8 +134,10 @@ void main() {
 
     test('uses provided context', () {
       final context = MockContext();
+      mock.when(() => context.nameFor(mock.any())).thenReturn('Test-Reaction');
+
       autorun((_) {}, context: context);
-      verify(context.runReactions());
+      mock.verify(() => context.runReactions());
     });
 
     test('can be disposed inside the tracking function', () {