+++ mobx/test/intercept_test.dart
@@ -1,8 +1,5 @@
-// @todo pavanpodila: remove once Mockito is null-safe
-// @dart = 2.10
-
 import 'package:mobx/mobx.dart';
-import 'package:mockito/mockito.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 
 import 'shared_mocks.dart';
@@ -92,7 +89,7 @@ void main() {
         ..add((_) {})
         ..interceptChange(WillChangeNotification());
 
-      verify(context.untracked(any));
+      verify(() => context.untracked(any()));
     });
   });
 }